Update feedback loop for Bot
Updated a few mix packages and made required code
changes to support the updates.

Changed the food truck Twitter lookups to be an 
async_stream task so that I can check if any trucks
actually returned a useful tweet.  If none did,
give a response to the slack channel so that users
know at least the Bot did its job.

Also, respond directly to the user who interacted
with us.